+++ src/main/java/com/supermartijn642/scarecrowsterritory/ScarecrowsTerritoryClient.java
@@ -1,8 +1,6 @@
 package com.supermartijn642.scarecrowsterritory;
 
-import com.supermartijn642.core.registry.ClientRegistrationHandler;
 import net.fabricmc.api.ClientModInitializer;
-import net.minecraft.world.item.DyeColor;
 
 /**
  * Created 7/11/2020 by SuperMartijn642
@@ -11,11 +9,5 @@ public class ScarecrowsTerritoryClient implements ClientModInitializer {
 
     @Override
     public void onInitializeClient(){
-        ClientRegistrationHandler handler = ClientRegistrationHandler.get("scarecrowsterritory");
-        for(ScarecrowType type : ScarecrowType.values()){
-            for(DyeColor color : DyeColor.values()){
-                handler.registerBlockModelRenderType(() -> type.blocks.get(color), type::getRenderLayer);
-            }
-        }
     }
 }