+++ changelog.md
@@ -1,2 +1,2 @@
 ### Scarecrows' Territory 1.1.11
-- Initial release of Scarecrows' Territory for Minecraft 1.21.6 & 1.21.7
+- Initial release of Scarecrows' Territory for Minecraft 1.21.9 & 1.21.10