+++ changelog.md
@@ -1,2 +1,4 @@
+## Update to Minecraft 26.1
+
 ### Scarecrows' Territory 1.1.11
 - Initial release of Scarecrows' Territory for Minecraft 1.21.11